requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

# 基础框架 (强制对齐官方 CUDA 12.8 推理要求)
torch==2.8.0+cu128
torchvision==0.23.0+cu128
torchaudio==2.8.0+cu128

# UI 与后端框架
gradio>=4.0.0
python-dotenv
pandas>=2.0.0
huggingface_hub>=0.20.0
omegaconf>=2.3.0
ffmpeg-python

# 音频处理与模型推理
audio-separator[gpu]>=0.17.0
pyannote.audio>=3.1.1
faster-whisper>=1.1.0
ctranslate2>=4.5.0
nltk